+++ pkg/dds/CHANGELOG.md
@@ -1,6 +1,7 @@
 # 2.9.1
 - [DAP] `runInTerminal` requests are now sent after first responding to the `launchRequest`.
 - [DAP] Skipped tests are now marked with `!` instead of `✓` in `Output` events.
+- [DAP] Implemented `pause` request.
 - Fixed a bug that was preventing clients from receiving `IsolateReload` events
   (see https://dartbug.com/49491).